ENH: Threshold values should be of the same type as the input pixel value.
SVN-Revision: 253

+++ Sandbox/Testing/itkLungWallFeatureGeneratorTest1.cxx
@@ -76,7 +76,7 @@ int main( int argc, char * argv [] )
 
   if( argc > 3 )
     {
-    featureGenerator->SetLungThreshold( atof( argv[3] ) );
+    featureGenerator->SetLungThreshold( atoi( argv[3] ) );
     }
